# mingldingl_engine/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "MinglDingl.Engine",
    "version": "1.0"
  },
  "paths": {
...
    "/scores/me": {
      "get": {
        "tags": [
          "Scores"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScoreResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/scores/me/detail": {
      "get": {
        "tags": [
          "Scores"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScoreDetailResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/scores/me/notifications/ack": {
      "post": {
        "tags": [
          "Scores"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AckNotificationDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AckNotificationDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AckNotificationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AckNotificationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/scores/me/history": {
      "get": {
        "tags": [
          "Scores"
        ],
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "cursorId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScoreHistoryResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/scores/leaderboard": {
      "get": {
        "tags": [
          "Scores"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeaderboardResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/scores/daily-login": {
      "post": {
        "tags": [
          "Scores"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyLoginResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AckNotificationDto": {
        "type": "object",
        "properties": {
          "kind": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AckNotificationResponse": {
        "type": "object",
        "properties": {
          "acknowledged": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "DailyLoginResponse": {
        "type": "object",
        "properties": {
          "awarded": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "currentStreak": {
            "type": "integer",
            "format": "int32"
          },
          "longestStreak": {
            "type": "integer",
            "format": "int32"
          },
          "streakBonusAwarded": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "DroppedItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "nameKey": {
            "type": "string",
            "nullable": true
          },
          "rarity": {
            "type": "string",
            "nullable": true
          },
          "itemType": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "nullable": true
          },
          "code": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "LeaderboardEntryDto": {
        "type": "object",
        "properties": {
          "rank": {
            "type": "integer",
            "format": "int32"
          },
          "gemTier": {
            "type": "string",
            "nullable": true
          },
          "score": {
            "type": "integer",
            "format": "int32"
          },
          "isCurrentUser": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "LeaderboardResponse": {
        "type": "object",
        "properties": {
          "city": {
            "type": "string",
            "nullable": true
          },
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LeaderboardEntryDto"
            },
            "nullable": true
          },
          "myRank": {
            "type": "integer",
            "format": "int32"
          },
          "hearthboundPairs": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "ScoreDetailResponse": {
        "type": "object",
        "properties": {
          "totalScore": {
            "type": "integer",
            "format": "int32"
          },
          "gemTier": {
            "type": "string",
            "nullable": true
          },
          "reputationScore": {
            "type": "number",
            "format": "double"
          },
          "currentStreak": {
            "type": "integer",
            "format": "int32"
          },
          "longestStreak": {
            "type": "integer",
            "format": "int32"
          },
          "tierIndex": {
            "type": "integer",
            "format": "int32"
          },
          "tierBonus": {
            "type": "integer",
            "format": "int32"
          },
          "nextTier": {
            "type": "string",
            "nullable": true
          },
          "nextTierThreshold": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "progressPct": {
            "type": "number",
            "format": "double"
          },
          "dailyMatchBudget": {
            "type": "integer",
            "format": "int32"
          },
          "pendingReferralReward": {
            "$ref": "#/components/schemas/DroppedItem"
          },
          "pendingShipReward": {
            "$ref": "#/components/schemas/DroppedItem"
          },
          "threadsSparked": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "ScoreEventDto": {
        "type": "object",
        "properties": {
          "eventType": {
            "type": "string",
            "nullable": true
          },
          "delta": {
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "ScoreHistoryResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScoreEventDto"
            },
            "nullable": true
          },
          "nextCursor": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "nextCursorId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ScoreResponse": {
        "type": "object",
        "properties": {
          "totalScore": {
            "type": "integer",
            "format": "int32"
          },
          "gemTier": {
            "type": "string",
            "nullable": true
          },
          "reputationScore": {
            "type": "number",
            "format": "double"
          },
          "dailyMatchBudget": {
            "type": "integer",
            "format": "int32"
          },
          "dailyMatchesUsed": {
            "type": "integer",
            "format": "int32"
          },
          "dailyMatchesRemaining": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}